requirements.txt
openai
flask
git+https://github.com/m-bain/whisperx.git
deepgram-sdk
python-dotenv
gcloud